feat(movie router): add DB req to find one movie
also move joi validation to middleware
