+++ routes/api/movies-router.js
@@ -1,59 +1,17 @@
 import express from 'express';
 
 const moviesRouter = express.Router(); // створює роутер
+
 import movieСontrolers from '../../controlers/movie_controlers.js';
+
+import { isValidId } from '../../middlewares/index.js';
+import { VadidateMovieData } from '../../middlewares/index.js';
 
 moviesRouter.get('/', movieСontrolers.getAllMovies);
 
-// moviesRouter.get('/:id', async (req, res, next) => {
-//   //всі динамічні параметри запиту зберігаються в змінній reg.params
-//   // console.log('reg.params : ', req.params );
-//   const { id } = req.params;
-//   // console.log('id: ', id);
-//   try {
-//     const result = await moviesService.getMovieById(id);
-//     // якщо id буде не правильний то moviesService.getMovieById поверне null (так працює база даних )null це не помилка а нам треба щоб коли id не вірний тобто коли база нічого не знайшла треба повертати помилку
-//     if (!result) {
-//       // res.status не перриває фннкцію в разі помилки тому треба ставити return щоб код далі не виконувавя коли відповілі з бази немає
-//       //   return   res.status(404).json({
-//       //         message:`обєкт з id:${id} не знайдено перевірте чи правильний id `
-//       //     })
+moviesRouter.get('/:id', isValidId, movieСontrolers.getMovieByID);
 
-//       //--variant2 ---
-//       // створити помилку самостійно і кикинути її якщо result = null
-//       // тоді код переривається і потрапляє в блок catch
-
-//       //         const error= new Error(`обєкт з id:${id} не знайдено перевірте чи правильний id `);
-//       // error.status=404;
-//       // throw error
-
-//       //--variant3  ---
-//       // створити допоміжну функцію helperі і перевикористовувати її
-//       throw HttpError(
-//         404,
-//         `обєкт з id:${id} не знайдено перевірте чи правильний id `
-//       );
-//     }
-
-//     res.json(result);
-//   } catch (error) {
-//     // res.status(500).json({
-//     //     message:"Server error"
-//     // })
-
-//     // деструктуризація обєкта error і присвоєння значення за замовчуванням якщо нічого не передали а помилка виникла статус буде 500
-//     //         const { status =500, message="Server error"}=error
-//     //           res.status( status).json({
-//     //            message,
-//     //    })
-
-//     //    якщо в параметр next передати error то express буде продовжувати виконувати код і шукати функцію обробник помилок
-//     // функцію обробник помилок  це функція яка має 4 аргументи (з документації express)
-//     next(error);
-//   }
-// });
-
-moviesRouter.post('/', movieСontrolers.addMovie);
+moviesRouter.post('/', VadidateMovieData, movieСontrolers.addMovie);
 
 // moviesRouter.put('/:id', async (req, res, next) => {
 //   // ПЕРЕВІРЯЄМ ТІЛО ЗАПИТУ ЧИ ВІДПОВІДАЄ СХЕМІ ВАЛАДАЦІЇ
